requirements.txt
# requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.25
alembic>=1.13.0
cryptography>=42.0.0
celery>=5.3.0
redis>=5.0.0
playwright>=1.40.0
boto3>=1.34.0
python-telegram-bot>=20.7
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
psycopg2-binary>=2.9.0
websockets>=11.0.0